move match to the proper location
